support float 64

+++ run_evaluation.py
@@ -76,7 +76,7 @@ if __name__ == "__main__":
         scene_name=scene_name,
         pretrained_ckpt=args.checkpoint,
         target_device=torch.device(args.device),
-        target_dtype=torch.float32 if args.dtype == "float32" else torch.float16,
+        target_dtype=torch.float32 if args.dtype == "float32" else torch.float64,
         ratio=0.5,
         use_rgb=scene_name == "plume_color_1",
         frame_start=frame_start,
